requirements.txt
matplotlib
numpy
cupy-cuda80
scikit-learn